Create a package extension for Krylov.jl
