+++ Project.toml
@@ -7,20 +7,24 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearOperators = "5c8ed15e-5a4c-59e4-a42b-c7e8811fb125"
 NLPModels = "a4795742-8479-5a88-8948-cc11e1c8c1a6"
 NLPModelsModifiers = "e01155f1-5c6f-4375-a9d8-616dd036575f"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatricesCOO = "fa32481b-f100-4b48-8dc8-c62f61b13870"
 
+[weakdeps]
+Krylov = "ba0b0d4f-ebba-5204-a429-3ac8c609bfb7"
+
+[extensions]
+LLSModelsKrylovExt = "Krylov"
+
 [compat]
+Krylov = "0.9.10"
 LinearOperators = "2"
-NLPModels = "0.18, 0.19, 0.20, 0.21"
-NLPModelsModifiers = "0.6, 0.7"
-Requires = "1"
-SparseMatricesCOO = "0.1, 0.2"
-julia = "^1.6.0"
+NLPModels = "0.21"
+NLPModelsModifiers = "0.7"
+SparseMatricesCOO = "0.2"
+julia = "1.10"
 
 [extras]
-Krylov = "ba0b0d4f-ebba-5204-a429-3ac8c609bfb7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NLPModelsTest = "7998695d-6960-4d3a-85c4-e1bceb8cd856"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
