+++ src/LLSModels.jl
@@ -5,6 +5,5 @@ using LinearOperators, NLPModels, NLPModelsModifiers, SparseMatricesCOO
 
 include("lls_model.jl")
 include("model-interaction.jl")
-include("solve_krylov.jl")
 
 end # module
